refactor: 게시글 작성자 소모임 소속 여부 메소드 수정 v2

+++ src/main/java/hobbiedo/board/application/ReplicaCommentServiceImpl.java
@@ -102,6 +102,7 @@ public class ReplicaCommentServiceImpl implements ReplicaCommentService {
 	// 댓글 작성자가 해당 소모임에 속한 회원인지 확인
 	private boolean isInCrew(String writerUuid, Long crewId) {
 
-		return replicaCrewRepository.existsByCrewIdAndMemberUuid(crewId, writerUuid).orElse(false);
+		return replicaCrewRepository.existsByCrewIdAndMemberUuid(crewId.longValue(), writerUuid)
+			.orElse(false);
 	}
 }
